Update to follow current changes
- Use 2018 for copyright date
- Use `Result` named constructors
- Use `TestAsset` namespace for barcode test classes
