+++ test/TestAsset/MyBarcode3.php
@@ -1,19 +1,19 @@
 <?php
 /**
  * @see       https://github.com/zendframework/zend-datavalidator for the canonical source repository
- * @copyright Copyright (c) 2017 Zend Technologies USA Inc. (http://www.zend.com)
+ * @copyright Copyright (c) 2018 Zend Technologies USA Inc. (http://www.zend.com)
  * @license   https://github.com/zendframework/zend-datavalidator/blob/master/LICENSE.md New BSD License
  */
 
-namespace ZendTest\DataValidator\Barcode;
+namespace ZendTest\DataValidator\TestAsset;
 
 use Zend\DataValidator\Barcode\AbstractAdapter;
 
-class MyBarcode2 extends AbstractAdapter
+class MyBarcode3 extends AbstractAdapter
 {
     public function __construct()
     {
-        $this->setLength([1, 3, 6]);
+        $this->setLength([1, 3, 6, -1]);
         $this->setCharacters(128);
         $this->setChecksum('_mod10');
     }
